[measureobjectintensitydistribution]: (4) Move measurement helper functions to cellprofiler/library/measurement.py, add types to function args
